+++ src/js/features/card-page-features/index.ts
@@ -1,0 +1,3 @@
+import TagMenu from "./tag-menu";
+
+export default [TagMenu];